feat(activity-settings): refactor activity settings to use a drawer component and improve UI interactions
