+++ libs/feature/course/browser/src/components/activity-card/activity-card.component.html
@@ -118,30 +118,14 @@
             <li nz-menu-item *ngIf="item.permissions.viewResource" [routerLink]="['/resources', item.resourceId]">
               Ouvrir la ressource associée
             </li>
-            <li nz-menu-item *ngIf="item.permissions.update" (click)="modal.open()">Paramètres</li>
+            <li nz-menu-item *ngIf="item.permissions.update" (click)="settingsDrawer.open()">Paramètres</li>
           </ul>
         </nz-dropdown-menu>
       </div>
     </mat-card-actions>
   </mat-card>
 </ng-container>
-<ui-modal-drawer size="large" [title]="item.title" [nzExtra]="drawerAction" #modal>
-  <ng-template>
-    <course-activity-settings
-      #settingsComponent
-      [(activity)]="item"
-      (activityChange)="onActivityChange($event)"
-      (saveRequested)="onSaveRequested()"
-    />
-  </ng-template>
-</ui-modal-drawer>
-
-<ng-template #drawerAction>
-  <button nz-button nzType="primary" nzSize="default" (click)="saveSettings()" class="save-btn">
-    <i nz-icon nzType="save" nzTheme="outline"></i>
-    Enregistrer les modifications
-  </button>
-</ng-template>
+<course-activity-settings-drawer [(activity)]="item" (activityChange)="onActivityChange($event)" #settingsDrawer />
 <ng-template #progressionTemplate>
   <div class="progression">
     <span class="progression-value">{{ item.progression }}</span>
