+++ libs/feature/course/browser/src/components/activity-card/activity-card.component.ts
@@ -8,9 +8,7 @@ import {
   OnDestroy,
   OnInit,
   CUSTOM_ELEMENTS_SCHEMA,
-  ViewChild,
 } from '@angular/core'
-import { NzDrawerModule } from 'ng-zorro-antd/drawer'
 import { RouterModule } from '@angular/router'
 import { firstValueFrom, Subscription } from 'rxjs'
 
@@ -23,17 +21,15 @@ import { NzIconModule } from 'ng-zorro-antd/icon'
 import { NzProgressModule } from 'ng-zorro-antd/progress'
 import { NzToolTipModule } from 'ng-zorro-antd/tooltip'
 import { NzDropDownModule } from 'ng-zorro-antd/dropdown'
-import { NzModalComponent, NzModalService } from 'ng-zorro-antd/modal'
 
 import { MatIconModule } from '@angular/material/icon'
 import { Activity } from '@platon/feature/course/common'
-import { UiModalDrawerComponent } from '@platon/shared/ui'
 import { ThemeService } from '@platon/core/browser'
 import { ResourceService } from '@platon/feature/resource/browser'
 import { CoursePipesModule } from '../../pipes'
-import { CourseActivitySettingsComponent } from '../activity-settings/activity-settings.component'
 import { CourseItemComponent } from '../course-item/course-item.component'
 import { CsvDownloadButtonComponent } from '../csv-download-button/csv-download-button.component'
+import { CourseActivitySettingsDrawerComponent } from '../activity-settings-drawer/activity-settings-drawer.component'
 import { CourseService } from '../../api/course.service'
 
 @Component({
@@ -62,10 +58,7 @@ import { CourseService } from '../../api/course.service'
     CourseItemComponent,
     CsvDownloadButtonComponent,
 
-    UiModalDrawerComponent,
-    NzDrawerModule,
-    CourseActivitySettingsComponent,
-    NzModalComponent,
+    CourseActivitySettingsDrawerComponent,
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
@@ -77,12 +70,6 @@ export class CourseActivityCardComponent implements OnInit, OnDestroy {
   private themeSubscription?: Subscription
 
   @Input() item!: Activity
-  @ViewChild('settingsComponent') settingsComponent?: CourseActivitySettingsComponent
-  @ViewChild('modal') modal?: UiModalDrawerComponent
-
-  protected showAccessWarningModal = false
-
-  constructor(private modalWarning: NzModalService) {}
 
   async ngOnInit(): Promise<void> {
     this.themeSubscription = this.themeService.themeChange.subscribe(() => {
@@ -140,97 +127,6 @@ export class CourseActivityCardComponent implements OnInit, OnDestroy {
 
   get editorUrl(): string {
     return this.resourceService.editorUrl(this.item.resourceId, 'latest')
-  }
-
-  protected async saveSettings(): Promise<void> {
-    if (!this.settingsComponent) {
-      return
-    }
-
-    const hasAccessPeriods = this.settingsComponent.accessPeriodsLength > 0
-    const hasOthersRule = this.settingsComponent.hasOthersRule()
-
-    if (hasAccessPeriods && !hasOthersRule) {
-      this.showAccessWarning()
-      return
-    }
-
-    await this.performSave()
-  }
-
-  private showAccessWarning(): void {
-    this.modalWarning.confirm({
-      nzTitle: "Configuration d'accès incomplète",
-      nzContent: this.createWarningContent(),
-      nzWidth: 600,
-      nzOkText: "Confirmer l'accès restreint",
-      nzOkType: 'primary',
-      nzOkDanger: true,
-      nzCancelText: 'Modifier les périodes',
-      nzOnOk: async () => {
-        await this.performSave()
-      },
-      nzOnCancel: () => {
-        // Ne rien faire, rester dans les paramètres
-      },
-    })
-  }
-
-  private createWarningContent(): string {
-    return `
-      <div class="access-warning-content">
-        <div style="display: flex; align-items: flex-start; gap: 16px; margin-bottom: 20px; padding: 16px; background: var(--brand-background-hover); border-radius: 8px; border-left: 4px solid var(--brand-color-primary);">
-          <div>
-            <h4 style="margin: 0 0 8px 0; color: var(--brand-text-primary);">Certains étudiants n'auront pas accès à cette activité</h4>
-          </div>
-        </div>
-
-        <p style="color: var(--brand-text-primary); margin: 16px 0 8px 0;"><strong>Situation actuelle :</strong></p>
-        <div style="background: var(--brand-background-components); border: 1px solid var(--brand-border-color); border-left: 4px solid var(--brand-color-secondary); border-radius: 6px; padding: 16px; margin: 12px 0 20px 0;">
-          Vous avez créé des périodes d'accès spécifiques, mais aucune période "Tous les autres".
-        </div>
-
-        <p style="color: var(--brand-text-primary); margin: 16px 0 8px 0;"><strong>Conséquences :</strong></p>
-        <div style="background: var(--brand-pastel-red); border: 1px solid var(--brand-border-color-light); border-left: 4px solid var(--brand-text-error); border-radius: 6px; padding: 16px; margin: 12px 0 20px 0;">
-          <ul style="margin: 0; padding-left: 20px;">
-            <li style="margin-bottom: 8px;">Seuls les étudiants dans vos groupes/listes spécifiés auront accès</li>
-            <li>Les nouveaux étudiants n'y auront pas accès automatiquement</li>
-          </ul>
-        </div>
-
-        <div style="background: var(--brand-pastel-green); border: 1px solid var(--brand-border-color); border-left: 4px solid var(--brand-color-secondary); border-radius: 6px; padding: 16px; margin: 20px 0 0 0;">
-          <strong style="color: var(--brand-color-secondary);">Recommandation :</strong>
-          Ajoutez une période d'accès avec le type "Tous les autres" si vous voulez que tous les étudiants du cours aient accès à cette activité.
-        </div>
-      </div>
-    `
-  }
-
-  private async performSave(): Promise<void> {
-    if (!this.settingsComponent) {
-      return
-    }
-
-    try {
-      await this.settingsComponent.update()
-      this.cdr.markForCheck()
-    } catch (error) {
-      console.error('Erreur lors de la sauvegarde:', error)
-    }
-  }
-
-  closeAccessWarningModal(): void {
-    this.showAccessWarningModal = false
-    this.cdr.markForCheck()
-  }
-
-  async confirmRestrictedSave(): Promise<void> {
-    this.showAccessWarningModal = false
-    await this.performSave()
-  }
-
-  protected onSaveRequested(): void {
-    this.modal?.close()
   }
 
   protected onActivityChange(activity: Activity): void {
