+++ libs/feature/course/browser/src/components/activity-settings-drawer/activity-settings-drawer.component.html
@@ -1,0 +1,17 @@
+<ui-modal-drawer size="large" [title]="localActivity()?.title || ''" [nzExtra]="drawerAction" #modal>
+  <ng-template>
+    <course-activity-settings
+      #settingsComponent
+      [activity]="localActivity()!"
+      (activityChange)="onActivityChange($event)"
+      (saveRequested)="onSaveRequested()"
+    />
+  </ng-template>
+</ui-modal-drawer>
+
+<ng-template #drawerAction>
+  <button nz-button nzType="primary" nzSize="default" (click)="saveSettings()" class="save-btn">
+    <i nz-icon nzType="save" nzTheme="outline"></i>
+    Enregistrer les modifications
+  </button>
+</ng-template>
