+++ libs/feature/course/browser/src/components/activity-settings/activity-settings.component.html
@@ -84,7 +84,7 @@
                 <span
                   >Tous les membres du cours ont accès à cette activité aux dates indiquées ci-dessus.<br />
                   Cliquez sur une date si vous souhaitez la changer.
-                  <b>N'oubliez pas d'enregister vos changement.</b></span
+                  <b>N'oubliez pas d'enregistrer vos changement.</b></span
                 >
               </div>
             </div>
