+++ libs/feature/course/browser/src/components/index.ts
@@ -15,3 +15,4 @@ export * from './csv-download-button/csv-download-button.component'
 export * from './course-member-search-modal/course-member-search-modal.component'
 export * from './course-group-select/course-group-select.component'
 export * from './activity-settings/activity-settings.component'
+export * from './activity-settings-drawer/activity-settings-drawer.component'
